+++ .eslintrc.js
@@ -13,6 +13,7 @@ module.exports = {
   extends: ["eslint:recommended", "prettier"],
   rules: {
     "no-unused-vars": ["error", { argsIgnorePattern: "^_" }],
+    "require-yield": "off",
     "prettier/prettier": "error"
   }
 };